Set guest_token cookie for guest orders
Something is blowing away this cookie through all of the redirects
that are happening and a new one gets assigned.

So by restoring this cookie for guest orders we can ensure that the
user is not redirected to the login page for their cookie guest
token not matching the one on the order.