Don't authorise any payments as "recurring"
This commit reverts some major changes made in an older commit:
075a625c33

The commit author explains that all payments would be authorised as
recurring payments because, otherwise, the tokens Adyen would send to
Solidus could expire before if the customer didn't explicitly complete
the "confirm" at checkout within one hour of the authorisation.

For our fork of this integration, this is not an issue: our "payment"
checkout step is also the "confirm" step. So there's no opportunity
for tokens to expire between the "payment" and "confirm" steps of
checkout.

---

Unfortunately, we're unable to get specs passing on CI. CI had never
been configured for this fork, so we don't have a record of when the
test suite stopped passing consistently. In my local development
environment, I've been unable to to get the suite passing.

This means that we'll have to rely on the applications's feature test
suite to test these changes.