Make specs pass with non-recurring payments
Previously, we made change to our authorisation flow, in that we no
longer authorise payments as "recurring". See commit
`ece334f6edef0cc0a37e8e9ead074abb4026294f` for more details on why we
did this. In that commit, the author mentions that specs are not
passing, and this change helps with that (a bit).