Fix firstname/lastname deprecation warnings (v3.0)
Although Adyen does not process HPP transactions anymore, rather than
spending the time to safely delete this code, I am chosing to move
quickly and just fix this code so it is compliant with the Solidus 3.0
upgrade.

Instead of firstname/lastname, Spree::Address only supports `name` as of
Solidus v3.0.