Fix refunds aka "credit" bug
Refunds were reported as non-functional on Sirdar today with the
following error:

NoMethodError
undefined method `payment' for #<Spree::Payment id: 248826...

Not really sure how this bug was introduced as we basically just renamed
the variable from options to gateway_options in our previous change to
support Ruby 3.0 via argument restructuring.

The diff looked like this:

-- currency ||= options[:originator].payment.currency
++ currency ||= gateway_options[:originator].payment.currency

Anyway it works again if you don't try to fetch the payment off the
payment, so that is the change I make here.

I also fix what is a real acutal bug where we would never have merged
the `:additional_data` from the gateway_options due to missing a step in
the rename.