Ignore failed capture events if they are dupes
Ok so technically they aren't dupes, because they are separate events
with separate webhook event ids (pspReferences) but they pretain to the
same authorize event/payment, so they are not supposed to exist.

We are getting them because we are authorizing twice, and then recieving
an equal number of capture events. It's a weird bug but the authorize
events are counted as duplicate events because they have the same
pspReference (the original payment pspRef which is created upon
authorize inside of Adyen), sucess code (true), and event type
(AUTHORIZE)).

So our logs (correctly) chuck out the second dupe authorize event. They
don't do this with the second capture event because they aren't true
dupes (the capture events have different success codes, first is true
and second is false since we ran out of money, and they also have
different pspReference/webhook event ids).

With this context out of the way, our intention with this fix is to add
logic to IGNORE the second capture event. Our fix here will only ignore
an event where:

- It's FAILED
- It's a CAPTURE event
- The payment already has an associated Spree::PaymentCaptureEvent
  (indicating that this incoming capture event is not the first)

The ideal fix would be to not end up sending two authorize events and
therefore to never receive two capture events, but we weren't able to
figure out the source of the issue, so this is still a #TODO.

Co-authored-by: benjamin wil <benjamin@super.gd>