Add cse_library_location_ingegrity preference
The integrity SHA is going to be required soon for CSE integrations.
For each legacy credit card payment method we need to include this
new information in the payment method preferences. This is going to
be used when generating the script tag on the FE for loading the
appropriate CSE library, and allows to verify the integrity of the
script. The integrity code is available in the Adyen admin area for
each payment method.